requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
psycopg2-binary==2.9.9
resend==0.8.0
python-multipart==0.0.6
twilio==8.11.0
flask==3.0.0
